Reading/learning about git
## Git
Git is a tool that acts as online storage that stores data files such as codes.
Its benefits comes from its ability to be accessed from different computers and different locations. Consequently, git provides protection to data because you will not lose your data if your computer crushing or something happens to it.
In addition to that git keeps track of changes we make to our programs and we are able to go back to a previous version if needed. Plus different people can work or contribute to the same code and the changes are incremented. Each person can import the files to their local machine, make changes and then update the git files.
### So benefits of git can be summarized as:
-	Provides a way to track changes
-	Protection against data loss 
-	Decentralized access to the data
How does get work:
In order to use git we have to install and configure it on our local computer.
We use the command line to manipulate data such as creating files, updating files etc.
We have to clone the project/repository using clone command with a link to  the git server. This enables us to have an exact copy of that project (repository). 
Some of the most used commands are:
-	cd (change directory) example cd somedirectory
-	add (adds file) git add somefile
-	commit ( commit changes) example git commit –m “some message”
-	push (pushes updates to git) using the  git push origin main 
-	status (checks the state of the files) command: git status


[Back to Home](README.md)